relocate `coreutils/zip.go` to `goutils/zipu`
